Unfinished prop modified flag
