+++ Cerulean.Components/Graphical/Label.cs
@@ -1,10 +1,12 @@
 ﻿
 using Cerulean.Common;
+using Cerulean.Core;
 
 namespace Cerulean.Components
 {
     public class Label : Component, ISized
     {
+        private Size? _oldClientArea = null;
         public Size? Size { get; set; } = null;
         public int? HintW { get; set; }
         public int? HintH { get; set; }
@@ -27,6 +29,11 @@ namespace Cerulean.Components
                 CallHook(this, EventHook.BeforeUpdate, window, clientArea);
 
             ClientArea = Size ?? clientArea;
+
+            if (_oldClientArea != ClientArea && window is Window ceruleanWindow)
+                ceruleanWindow.FlagForRedraw();
+
+            _oldClientArea = ClientArea;
 
             if (window is not null)
                 CallHook(this, EventHook.AfterUpdate, window, clientArea);
